### 2024-10-23 #7338 @kthyng
@fgerick
* Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html
