### 2024-11-05 bc05f2b
shorten joss paper
