+++ paper/paper.md
@@ -48,25 +48,7 @@ The model code base is tested against mode solutions from the scientific literat
 
 # Theoretical background and implementation details
 
-In non-linear form, the momentum equation of the incompressible fluid and the induction equation, are written as
-$$
-\begin{aligned}
-	\frac{\partial \mathbf{U}}{\partial t} + \left(\boldsymbol{\nabla}\times\mathbf{U}\right)\times\mathbf{U} + 2\boldsymbol{\Omega}\times\mathbf{U} &= -\frac{1}{\rho}\nabla P + \frac{1}{\rho\mu_0} \left(\boldsymbol{\nabla}\times\mathbf{B}\right)\times\mathbf{B} + \nu\boldsymbol{\nabla}^2\mathbf{U} + \mathbf{F}, \\
-	\frac{\partial \mathbf{B}}{\partial t} &= \boldsymbol{\nabla}\times\left(\mathbf{U}\times\mathbf{B}\right) + \eta \boldsymbol{\nabla}^2\mathbf{B},
-\end{aligned}
-$$
-with $\mathbf{U}$ the velocity, $\mathbf{B}$ the magnetic field, $\boldsymbol{\Omega}$ the rotation axis, $\rho$ the fluid density, $P$ the reduced hydrodynamic pressure, $\mu_0$ the magnetic permeability of free space, $\nu$ the kinematic viscosity, $\mathbf{F}$ some additional body force, and $\eta$ the magnetic diffusivitiy.
-
-In order to compute modal solutions, the velocity, magnetic and pressure fields a linearized, so that 
-$$
-\begin{aligned}
-	\mathbf{U}(\mathbf{r},t) & = \mathbf{U}_0(\mathbf{r})+ \mathbf{u}(\mathbf{r}) e^{\lambda t}, \\
-	\mathbf{B}(\mathbf{r},t) & = \mathbf{B}_0(\mathbf{r})+ \mathbf{b}(\mathbf{r}) e^{\lambda t}, \\
-	P(\mathbf{r},t)   & = P_0(\mathbf{r})+ p(\mathbf{r}) e^{\lambda t}.
-\end{aligned}
-$$
-with $\lambda=-\sigma+\mathrm{i}\omega$, with $\sigma$ the damping rate and $\omega$ the frequency of the oscillatory perturbation to the steady background.
-Removing the steady part and neglecting higher order terms, the linearized momentum and induction equations then read
+In order to compute modal solutions, we consider the linearized momentum equation of the incompressible fluid and the linearized induction equation
 $$
 \begin{aligned}
 	\lambda\mathbf{u} =& -\left(\boldsymbol{\nabla}\times\mathbf{u}\right)\times\mathbf{U}_0- \left(\boldsymbol{\nabla}\times\mathbf{U}_0\right)\times\mathbf{u} -2\Omega\mathbf{e}_z\times\mathbf{u} - \frac{1}{\rho}\nabla p\\
@@ -74,12 +56,9 @@ $$
 	\lambda\mathbf{b} =& \boldsymbol{\nabla}\times\left(\mathbf{U}_0\times\mathbf{b}\right) + \boldsymbol{\nabla}\times\left(\mathbf{u}\times\mathbf{B}_0\right) + \eta \boldsymbol{\nabla}^2\mathbf{b}.
 \end{aligned}
 $$
[…]
