### 2024-11-05 #7338 @fgerick
I have shortened the paper to meet the required word count (I used a markdown wordcounter to get to 944 words excluding the header and references). Please let me know if additional shortening is needed. Thank you and best regards.
