### 2024-11-18 #7338 @diehlpk
> @editorialbot invite @diehlpk as editor

I can take that submission but can only start after Thanksgiving due to travel.
