### 2024-12-12 #7338 @jsoishi
> Hi @jsoishi do you have time to review this paper?
Unfortunately, I am fully booked for the remainder of the year.
