### 2024-12-16 #7593 @editorialbot
[REVIEW]: Limace.jl: A Julia package to compute hydromagnetic modes in spherical domains
**Submitting author:** @fgerick (<a href="http://orcid.org/0000-0001-9924-0562">Felix Gerick</a>)
**Repository:** https://github.com/fgerick/Limace.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
