### 2025-01-10 e8714c2
add SHTns grid option to discretization
