### 2025-03-13 fgerick/Limace.jl#13 @milankl
Add a progress meter?
Apologies for the delay in this, the last weeks and months were hectic with a lot of additional responsibilities in my job. I'm opening up a few issues now.

```julia
julia> @time LHS, RHS, u, b = assemble(N, Le, Lu, B0);
 51.393637 seconds (204.33 k allocations: 68.522 MiB, 0.43% gc time)
```

This step in the documenatation takes a considerable amount of time. I would suggest to include some progress meter here so that a user has some feedback of what's going on here? Putting `@time` in front of each line in `assemble` yields

| What | Time |
| - | -|
| Inviscid | 0.000001 seconds (2 allocations: 160 bytes) |
| Insulating | 0.000000 seconds (2 allocations: 160 bytes) |
| LHS | 0.018341 seconds (39.85 k allocations: 2.827 MiB, 82.45% compilation time) |
| RHSc | 0.006216 seconds (250 allocations: 4.062 MiB) |
| RHSl | 36.372186 seconds (96.23 k allocations: 29.410 MiB, 0.27% gc time) |
| RHSi | 14.708092 seconds (68.60 k allocations: 22.502 MiB, 0.01% gc time, 0.05% compilation time) |
| RHSd |  0.000255 seconds (77 allocations: 904.328 KiB) |
| RHS | 0.008514 seconds (228 allocations: 8.874 MiB, 19.38% gc time) |

given the bulk is with `_threaded` functions it sounds like you could put some version of the ProgressMeter.jl around it? Also how do there ~50s scale with `N`? How much larger `N`s would a user want realistically want to use?
