### 2025-03-13 fgerick/Limace.jl#15 @milankl
User interface
It seems to me that to use this software there are conceptually 4 steps that one has to do

1. Choose a basis, resolution
2. Choose the terms in the equations / forces
3. Assemble the problem
4. Solve the problem

This is (despite other terminology) super similar to what we do in Oceananigans and SpeedyWeather. There it's about 1) creating a "grid", 2) defining "components" (which are often terms in the equations but can also be certain numerics/solvers used, or computing analysis on the fly, storing output etc, 3) constructing a model from these components and initializing it and 4) run a simulation. If you agree with this high-level description of the user interface / workflow then I would suggest to also reflect that in the code interfaces so that the structure is clearer to a user. For example, (you need to know how to name that exactly)

```julia
# 1. define a basis
basis = Inviscid(N=8)    # or Insulating, ...

# 2. define the terms/components/forces
lorentz = Lorentz(basis, ...)
induction = Induction(basis, ...)
coriolis = Coriolis(basis, Le=5e-4)

# 3. construct a model / assemble it
model = TorsionalAlfven(basis; lorentz, induction, coriolis) # not specifying diffusion uses defaults
problem = assemble!(model)

# 4. solve it!
solve!(problem)
```

And one could then use `InternationalModes` instead of `TorsionalAlfven` etc... or do `coriolis = NoCoriolis(basis)` etc. But overall this would make it much easier for a user to create their own problems and solve them! Let me know what you think!

https://github.com/openjournals/joss-reviews/issues/7593
