### 2025-03-13 fgerick/Limace.jl#15 @milankl
Given you talk about [Forces](https://fgerick.github.io/Limace.jl/dev/forces/) in the documentation you can also think of `forces` being a dictionary, so that one could do

```julia
forces = Forces()
add!(forces, Coriolis(basis))
add!(forces, Diffusion(basis, viscosity=1))
model = TorsionalAlfven(basis, forces)
```

and in `assemble!(model)` for `force in forces` is then evaluated and adds sparse matrices to the `RHS` sparse matrix?
