### 2025-03-13 fgerick/Limace.jl#15 @milankl
Also given that you talk about [Postprocessing](https://fgerick.github.io/Limace.jl/dev/processing/) in the documentation I'm wondering whether you can wrap that functionality into nicely reusable `spectrum(problem)` that can be called to analyse/visualise a solution to a problem after you ran `solve!(problem)` or maybe you want to separate solution from the problem as independent objects, like `solution = solve!(problem); spectrum(solution)` ? Not sure what's best in your case here ...
