### 2025-03-13 fgerick/Limace.jl#17 @milankl
JOSS Review
This is an issue for the general review of https://github.com/openjournals/joss-reviews/issues/7593 with the specific issues in

- #11
- #12
- #13
- #14
- #15
- #16
- #18

Limace.jl is a software package to solve the linear magnetohydrodynamics (MHD) on the sphere. The user can choose between different bases (=boundary conditions for the variables velocity $$\mathbf{U}, \mathbf{B}$$), set up forces, assemble these into sparse matrices that can sequentially be solved to solve a problem of the linear MHD. The code is written in a modular way such that different parts of Limace's functionality can be reused to solve different types of problems. Examples are provided for the Malkus modes, intertial modes and the torsional Alfvén modes, including analysis and visualisation. The solver is based on the [SHTns.jl](https://github.com/fgerick/SHTns.jl) wrapper of a C library of the same name, providing highly performant spherical harmonic transforms. The package uses unit tests, has a documentation which also uses Literate.jl (nice!) and provides some theoretical background on the equations being solved.

The scope of this software is very relevant, I am not aware of other MHD solvers for spherical domains in Julia. There is https://github.com/MHDFlows/MHDFlows.jl but this seems to be in Cartesian coordinates only. I love the flexibility of the functionality, that is share between setting up different problems instead of providing monolithic `solve_internial_models()` interfaces. However, I believe to structure this better and to make the usage of this package more intuitive it would be great to provide some more abstraction into these workflows as outlined in #15. I've raised two issues around performance #11 and #12 and for larger problems a user should get some information about run time #13. Finally there seems to be some reproducibility issues with the docs that can probably easily be resolved #14, and extended #16. Otherwise kudos to the author for providing some nice software to simulate magnetohydrodynamics to the Julia community. At the moment the package is not registered, but I believe acceptance for JOSS would and should register this as at least v0.1.

Please leave this issue open until the issues listed above are closed.
