### 2025-03-13 fgerick/Limace.jl#15 @milankl
Note that you do provide these high-level steps here https://fgerick.github.io/Limace.jl/dev/#Getting-started 😉
