### 2025-03-13 #7593 @milankl
Apologies for the delay @diehlpk, I just finished the first round of my review and linked all issues. All further information is in there and I'm more than happy to accept this paper for publication in JOSS upon resolving those issues!
