### 2025-03-26 7610359
norm torsional mode for reproducible plot
