+++ dev/assets/documenter.js
@@ -240,6 +240,20 @@ $(document).ready(function () {
       type: "click",
       noToggleAnimation: true,
     });
+
+    setTimeout(function () {
+      if (window.location.hash) {
+        const targetId = window.location.hash.substring(1);
+        const targetElement = document.getElementById(targetId);
+
+        if (targetElement) {
+          targetElement.scrollIntoView({
+            behavior: "smooth",
+            block: "center",
+          });
+        }
+      }
+    }, 100);
   }
 });
 
@@ -613,175 +627,249 @@ function worker_function(documenterSearchIndex, documenterBaseURL, filters) {
   };
 }
 
-// `worker = Threads.@spawn worker_function(documenterSearchIndex)`, but in JavaScript!
-const filters = [
-  ...new Set(documenterSearchIndex["docs"].map((x) => x.category)),
-];
-const worker_str =
-  "(" +
-  worker_function.toString() +
-  ")(" +
-  JSON.stringify(documenterSearchIndex["docs"]) +
-  "," +
-  JSON.stringify(documenterBaseURL) +
-  "," +
-  JSON.stringify(filters) +
-  ")";
-const worker_blob = new Blob([worker_str], { type: "text/javascript" });
-const worker = new Worker(URL.createObjectURL(worker_blob));
-
 /////// SEARCH MAIN ///////
 
-// Whether the worker is currently handling a search. This is a boolean
-// as the worker only ever handles 1 or 0 searches at a time.
-var worker_is_running = false;
+function runSearchMainCode() {
+  // `worker = Threads.@spawn worker_function(documenterSearchIndex)`, but in JavaScript!
+  const filters = [
+    ...new Set(documenterSearchIndex["docs"].map((x) => x.category)),
+  ];
+  const worker_str =
+    "(" +
+    worker_function.toString() +
+    ")(" +
+    JSON.stringify(documenterSearchIndex["docs"]) +
+    "," +
+    JSON.stringify(documenterBaseURL) +
+    "," +
+    JSON.stringify(filters) +
+    ")";
+  const worker_blob = new Blob([worker_str], { type: "text/javascript" });
+  const worker = new Worker(URL.createObjectURL(worker_blob));
 
-// The last search text that was sent to the worker. This is used to determine
-// if the worker should be launched again when it reports back results.
-var last_search_text = "";
+  // Whether the worker is currently handling a search. This is a boolean
+  // as the worker only ever handles 1 or 0 searches at a time.
+  var worker_is_running = false;
 
-// The results of the last search. This, in combination with the state of the filters
-// in the DOM, is used compute the results to display on calls to update_search.
-var unfiltered_results = [];
+  // The last search text that was sent to the worker. This is used to determine
+  // if the worker should be launched again when it reports back results.
+  var last_search_text = "";
 
-// Which filter is currently selected
-var selected_filter = "";
+  // The results of the last search. This, in combination with the state of the filters
+  // in the DOM, is used compute the results to display on calls to update_search.
+  var unfiltered_results = [];
 
-$(document).on("input", ".documenter-search-input", function (event) {
-  if (!worker_is_running) {
-    launch_search();
-  }
-});
+  // Which filter is currently selected
+  var selected_filter = "";
 
-function launch_search() {
-  worker_is_running = true;
-  last_search_text = $(".documenter-search-input").val();
-  worker.postMessage(last_search_text);
-}
-
-worker.onmessage = function (e) {
-  if (last_search_text !== $(".documenter-search-input").val()) {
-    launch_search();
-  } else {
-    worker_is_running = false;
-  }
-
-  unfiltered_results = e.data;
-  update_search();
-};
-
-$(document).on("click", ".search-filter", function () {
[…]
