+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"forces/#Forces","page":"Forces","title":"Forces","text":"","category":"section"},{"location":"forces/","page":"Forces","title":"Forces","text":"Different forces included in Limace.jl. ","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"So far, there is no heat equation in Limace.jl, so the forces include only the mass-term (or Limace.inertial), the Coriolis force (Limace.coriolis), the Lorentz force (Limace.lorentz) and diffusion (Limace.diffusion).  In the induction equation we have the magnetic induction term nablatimesmathbfutimesmathbfB (Limace.induction). Some details for each of these forcings is given on the respective pages.","category":"page"},{"location":"forces/#Inertial","page":"Forces","title":"Inertial","text":"","category":"section"},{"location":"forces/","page":"Forces","title":"Forces","text":"int mathbfu_i^* cdot mathbfu_jmathrmdV","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"Limace.inertial","category":"page"},{"location":"forces/#Limace.inertial","page":"Forces","title":"Limace.inertial","text":"inertial(\n    b::Basis;\n    external\n) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}\n\n\nCompute the Galerkin projection matrix of basis b onto itself, i.e. the inner products.  Also known as the mass matrix.\n\n\n\n\n\n","category":"function"},{"location":"forces/#Coriolis","page":"Forces","title":"Coriolis","text":"","category":"section"},{"location":"forces/","page":"Forces","title":"Forces","text":"The necessary functions to compute Galerkin projections on the Coriolis term","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"int mathbfu_i^* cdot 2mathbfe_ztimesmathbfu_jmathrmdV","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"where mathbfu_ij are either poloidal or toroidal basis vectors.","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"Limace.coriolis","category":"page"},{"location":"forces/#Limace.coriolis","page":"Forces","title":"Limace.coriolis","text":"coriolis(\n    b;\n    Ω\n) -> SparseArrays.SparseMatrixCSC{ComplexF64, Int64}\n\n\nCompute the sparse Galerkin projection matrix, by projecting the basis b onto the Coriolis operator.\n\n\n\n\n\n","category":"function"},{"location":"forces/#Diffusion","page":"Forces","title":"Diffusion","text":"","category":"section"},{"location":"forces/","page":"Forces","title":"Forces","text":"int mathbfu_i^* cdot boldsymbolnabla^2mathbfu_jmathrmdV","category":"page"},{"location":"forces/","page":"Forces","title":"Forces","text":"Limace.diffusion","category":"page"},{"location":"forces/#Limace.diffusion","page":"Forces","title":"Limace.diffusion","text":"diffusion(\n    b::Basis;\n    external\n) -> SparseArrays.SparseMatrixCSC{Float64, Int64}\n\n\nCompute the Galerkin projection matrix of the basis b onto the vector Laplacian. When keyword
[…]
