### 2025-03-26 fgerick/Limace.jl#15 @fgerick
This is a big complicated task. I have thought about how to effectively implement the highest level API several times and I'm still not sure what the best way forward is here. The problem is to be generic, i.e. we want to be able to include several forcings and several ways of nondimensionalization (e.g. timescale is length-of-day or Alfvén time), and making it straightforward to adjust parameters without reassembling costly Galerkin projections. I'm afraid if I wrap everything into an opaque high-level assemble function, it will make it essentially harder for the user to go beyond what the examples does. I will think about this a bit more carefully, but it might take some more time.
