### 2025-04-04 13e4f7c
add malkus modes example
