+++ examples/malkusmodes.jl
@@ -1,0 +1,181 @@
+# # Malkus modes in the sphere
+#
+# [malkus_hydromagnetic_1967](@citet) presented analytical solutions of hydromagnetic modes in the full sphere, 
+# assuming a background magnetic field that is purely azimuthal (``\mathbf{B}_0 = s \mathbf{e}_z``).
+# Here, we will compute and visualize some of these modes numerically, using `Limace.jl`.
+#
+# The evolution equations of the velocity ``\mathbf{u}`` and magnetic field ``\mathbf{b}`` are given by
+#
+# ```math
+# \begin{align}
+# \frac{\partial\mathbf{u}}{\partial t} + \frac{2}{\mathrm{Le}} \mathbf{e}_z\times\mathbf{u} &= -\nabla p + \nabla\times\mathbf{B}_0\times\mathbf{b} + \nabla\times\mathbf{b}\times\mathbf{B}_0,\\
+# \frac{\partial\mathbf{b}}{\partial t} &= \nabla\times\left( \mathbf{u}\times\mathbf{B}_0\right),
+# \end{align}
+# ```
+#
+# satisfying ``\mathbf{u}\cdot\mathbf{n} = \mathbf{b}\cdot\mathbf{n} = 0`` at ``r=1`` (the surface of the sphere).
+#
+# By projecting this equation onto poloidal and toroidal basis vectors, we eliminate the pressure ``p``.
+#
+# Assuming ``\mathbf{u}(\mathbf{r},t) = \mathbf{u}(\mathbf{r}) \exp(\lambda t)`` (and analogously for ``\mathbf{b}``) , the system of equations reduces to an eigen problem
+#
+# ```math
+# \lambda \mathbf{B}\mathbf{x} = \mathbf{A}\mathbf{x},
+# ```
+#
+# where the eigen vector ``\mathbf{x}`` contain the spectral coefficients of the bases elements and
+#
+# ```math
+# \mathbf{B} = \begin{pmatrix} \mathbf{B}^u & 0 \\ 0 & \mathbf{B}^b \end{pmatrix},
+# ```
+#
+# with 
+# ```math
+# B^{[u,b]}_{ij} = \int [\mathbf{u},\mathbf{b}]_i \cdot [\mathbf{u},\mathbf{b}]_j\,\mathrm{d}V,
+# ```
+#
+# The right-hand side is given by
+# ```math
+# \mathbf{A} = \begin{pmatrix} \mathbf{A}^u & \mathbf{A}^{ub} \\ \mathbf{A}^{bu} & 0 \end{pmatrix},
+# ```
+#
+# with
+#
+# ```math
+# \begin{align}
+# A^u_{ij} &= \int \mathbf{u}_i \cdot \left(2\Omega\mathbf{e}_z\times\mathbf{u}_j\right)\,\mathrm{d}V,\\
+# A^{ub}_{ij} &= \int \mathbf{u}_i \cdot \left(\nabla\times\mathbf{B}_0\times\mathbf{b}_j + \nabla\times\mathbf{b}_i\times\mathbf{B}_0\right)\,\mathrm{d}V,\\
+# A^{bu}_{ij} &= \int \mathbf{b}_i \cdot \left(\nabla\times\mathbf{u}_j\times\mathbf{B}_0\right)\,\mathrm{d}V.
+# \end{align}
+# ```
+#
+#
+
+
+# ## Solve using Limace.jl
+# 
+# First, load the packages:
+
+using Limace
+using LinearAlgebra, SparseArrays, CairoMakie, GeoMakie
+using Limace.Discretization: spectospat
+CairoMakie.activate!(; type="png", px_per_unit=2)
+
+# Define a polynomial truncation degree `N`.
+
[…]
