+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"Braginsky, S. I. (1970). Torsional Magnetohydrodynamics Vibrations in the Earth's Core and Variations in Day Length. Geomagn. Aeron. 10, 3–12.\n\n\n\nBullard, E. C. and Gellman, H. (1954). Homogeneous Dynamos and Terrestrial Magnetism. Philosophical Transactions of the Royal Society of London. Series A, Mathematical and Physical Sciences 247, 213–278.\n\n\n\nChen, L.; Herreman, W.; Li, K.; Livermore, P. W.; Luo, J. W. and Jackson, A. (2018). The Optimal Kinematic Dynamo Driven by Steady Flows in a Sphere. Journal of Fluid Mechanics 839, 1–32.\n\n\n\nGerick, F. and Livermore, P. W. (2024). Interannual Magneto–Coriolis Modes and Their Sensitivity on the Magnetic Field within the Earth's Core. Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences 480, 20240184.\n\n\n\nIvers, D. J. and Phillips, C. G. (2008). Scalar and Vector Spherical Harmonic Spectral Equations of Rotating Magnetohydrodynamics. Geophysical Journal International 175, 955–974.\n\n\n\nLivermore, P. W. (2014). A Compendium of Galerkin Orthogonal Polynomials.\n\n\n\nZhang, K.; Earnshaw, P.; Liao, X. and Busse, F. H. (2001). On Inertial Waves in a Rotating Fluid Sphere. Journal of Fluid Mechanics 437, 103–119.\n\n\n\n","category":"page"},{"location":"processing/#Postprocessing","page":"Postprocessing","title":"Postprocessing","text":"","category":"section"},{"location":"processing/","page":"Postprocessing","title":"Postprocessing","text":"A non-exhaustive collection of spectral to spatial discretization and postprocessing routines.","category":"page"},{"location":"processing/","page":"Postprocessing","title":"Postprocessing","text":"Modules = [Limace.Discretization]","category":"page"},{"location":"processing/#Limace.Discretization.spectospat-Union{Tuple{Tr}, Tuple{T}, Tuple{TB}, Tuple{TU}, Tuple{Vector{T}, TU, TB, Tr, Int64, Int64}} where {TU<:Basis, TB<:Basis, T<:ComplexF64, Tr<:Union{Float64, AbstractVector{Float64}}}","page":"Postprocessing","title":"Limace.Discretization.spectospat","text":"spectospat(\n    coeffs::Array{T<:ComplexF64, 1},\n    u::Basis,\n    b::Basis,\n    r::Union{Float64, AbstractVector{Float64}},\n    nθ::Int64,\n    nϕ::Int64;\n    kwargs...\n) -> Tuple{Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Array{T, 3} where T<:ComplexF64, Vector{Float64}, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}\n\n\nTransform eigenvector containing spectral coefficients to two-dimensional vector fields at 1 x nθ x nϕ grid points at radius r. Returns ur,uθ,uϕ, br,bθ,bϕ, θ,ϕ.\n\n\n\n\n\n","category":"method"},{"location":"processing/#Limace.Discretization.spectospat-Union{Tuple{T}, Tuple{TB
[…]
