+++ dev/refs.bib
@@ -76,3 +76,14 @@
   volume = {10},
   pages = {3--12}
 }
+
+@article{malkus_hydromagnetic_1967,
+  title = {Hydromagnetic Planetary Waves},
+  author = {Malkus, Willem V R},
+  year = {1967},
+  journal = {Journal of Fluid Mechanics},
+  volume = {28},
+  number = {4},
+  pages = {793--802},
+  doi = {10.1017/S0022112067002447}
+}
