### 2025-04-07 4d34e52
change inviscid inertial to have diagonal with size
