### 2025-04-07 b2c8b60
change eigenproblem to eigen and add documentation
