+++ docs/src/eigen.md
@@ -1,0 +1,123 @@
+# Solving the eigenvalue problem
+
+The generalized eigenvalue problem is given as
+```math
+\lambda \mathbf{B}\mathbf{x} = \mathbf{A}\mathbf{x},
+```
+where ``\lambda`` is the eigenvalue and ``\mathbf{x}`` the corresponding (right) eigenvector.
+The linear operators (or matrices) ``\mathbf{A}`` and ``\mathbf{B}`` contain the Galerkin-projected operators of the system of equations.
+In the context of `Limace.jl`, these matrices are often referred to as `LHS` (_"left-hand-side"_) for ``\mathbf{B}``, and `RHS` (_"right-hand-side"_) for ``\mathbf{A}``.
+When ``\mathbf{B}`` is the unit matrix, the eigenvalue problem reduces to the so-called standard eigenvalue problem. This is for example the case when ``\mathbf{B}`` only contains the inner products of orthonormal bases.
+
+Solving the generalized eigenvalue problem is not a trivial task for matrices that are large (already at a dimension ``\sim 10`` this becomes difficult analytically).
+For different sizes of ``\mathbf{A}`` and ``\mathbf{B}``, we can/have to rely on different numerical methods to solve for eigensolutions ``(\lambda,\mathbf{x})``.
+
+## Dense vs. sparse solutions
+
+When the matrices ``\mathbf{A}`` and ``\mathbf{B}`` are small enough, it is possible to represent them as dense matrices and to compute the full (_dense_) spectrum of eigenvalues and eigenvectors. 
+_"Small enough"_ depends on the computer you are using, but for matrix sizes larger than ``10^4``, solving for all eigenvalues (and eigenvectors) is no longer feasible even on the average workstation. 
+Such matrix sizes are quickly approached for truncation degrees ``N\geq 40`` in `Limace.jl`, when no symmetry assumptions can be used (e.g. we cannot consider azimuthal degrees independently due to non-axisymmetry).
+
+Luckily, there are iterative methods to solve only for a subset of eigenvalues of the large sparse matrix system. 
+Several different methods exist, with one widely used method being the Arnoldi method, which relies on the orthonormalization of Krylov subspaces. 
+Through these methods we can compute a few eigenvalues at the extremes of the spectrum (e.g. the largest eigenvalue in absolute value), and also eigenvalues close to a target by shifting the spectrum.
+
+How to access these different eigen solvers is outlined in the following.
+
+## Dense eigenvalue solver
+
+The dense eigensolver routines provided through [LAPACK](https://www.netlib.org/lapack/) are available in Julia through the standard library `LinearAlgebra`:
+```julia
+using LinearAlgebra
+```
+
+Let us consider the torsional mode problem (see examples), which assembles the two matrices `LHS` and `RHS`, so that
+```julia
+using Limace
+
+N = 40
+Le = 5e-4
+Lu = 1/Le
+B0 = [BasisElement(Basis{Insulating}, Poloidal, (1,0,1),0.3), BasisElement(Basis{Insulating}, Poloidal, (2,0,1),0.7)]
+
+u = Inviscid(N; m=0)
+b = Insulating(N; m=0)
+LHS = blockdiag(sparse(Limace.inertial(u)), sparse(Limace.inertial(b)))
+RHSc = Limace.coriolis(u)/Le
+RHSl = sum(Limace.lorentz_threaded(u,b,B) for B in B0)
+RHSi = sum(Limace.induction_threaded(b,u,B) for B in B0)
+RHSd = Limace.diffusion(b)/Lu
[…]
