+++ src/Limace.jl
@@ -25,7 +25,6 @@ using .Discretization
 include("Eigen.jl")
 using .Eigen
 
-include("problem.jl")
 
 # forces
 include("forces/inertial.jl")
