### 2025-04-07 2dee240
fix eigs and test it
