+++ test/misc.jl
@@ -74,3 +74,13 @@ end
 	end
 
 end
+
+
+
+@testset "eigs" begin
+    N = 20
+    b = Limace.Inviscid(N)
+    RHS = Limace.coriolis(b)
+	max_eval = first(first(Limace.Eigen.eigs(RHS; nev=1)))
+	@test abs(max_eval) ≤ 2.0
+end
