+++ test/runtests.jl
@@ -6,10 +6,10 @@ using FunctionZeros
 using SparseArrays
 
 using Limace.Eigen: eigstarget
-
+using Limace.Bases: Sphere
 
 struct LJ22; end
-Limace.Bases.s(::Type{Basis{LJ22}}, V::Volume, l, m, n, r) = r^2 * (157 - 296r^2 + 143r^4) / (16 * sqrt(182 / 3))
+Limace.Bases.s(::Type{Basis{LJ22, Sphere}}, V::Sphere, l, m, n, r) = r^2 * (157 - 296r^2 + 143r^4) / (16 * sqrt(182 / 3))
 
 
 include("misc.jl")
