### 2025-04-14 73398af
lots of documentation, eigen to eigensolve, add solve limaceproblem
