### 2025-04-14 de2e29d
add sphere to unconstrained
