### 2025-04-15 fgerick/Limace.jl#13 @fgerick
This is actually more complicated than it seems. The loops in the assembly of the Galerkin matrices are not uniform and the number of elements to be computed is pretty difficult to estimate. I have not managed to get a useful version of ProgressMeter.jl running.
