### 2025-04-15 #7593 @fgerick
I have resolved all the issues raised by @milankl, with an exception of https://github.com/fgerick/Limace.jl/issues/13 , which is not straightforward to implement. Given that it's more of an additional convenient feature (show the progress of the computation), I don't think this is a crucial issue. Let me know if there are any other things to improve and I hope it is not to confusing for @jakelangham with the implemented changes.
