### 2025-04-19 c2a966f
add discretization tests
