+++ test/runtests.jl
@@ -12,6 +12,7 @@ struct LJ22; end
 Limace.Bases.s(::Type{Basis{LJ22, Sphere}}, V::Sphere, l, m, n, r) = r^2 * (157 - 296r^2 + 143r^4) / (16 * sqrt(182 / 3))
 
 
+include("discretize.jl")
 include("misc.jl")
 include("modes.jl")
 include("bases_nobc.jl")
