+++ src/Discretization.jl
@@ -11,15 +11,15 @@ using StaticArrays
 using SHTns
 
 
-function poloidal_discretize(::Type{Basis{T}}, V::Volume, l, m, n, r, θ, ϕ) where {T}
+function poloidal_discretize(::Type{Basis{T, Vol}}, V::Vol, l, m, n, r, θ, ϕ) where {T, Vol<:Volume}
     if r <= V.r1
-        dsrdr = ∂(r -> s(Basis{T}, V, l, m, n, r) * r, r)
-        ur = l * (l + 1) * s(Basis{T}, V, l, m, n, r) * ylm(l, m, θ, ϕ) / r
+        dsrdr = ∂(r -> s(Basis{T, Vol}, V, l, m, n, r) * r, r)
+        ur = l * (l + 1) * s(Basis{T, Vol}, V, l, m, n, r) * ylm(l, m, θ, ϕ) / r
         uθ = 1 / r * dsrdr * dylmdθ(l, m, θ, ϕ)
         uϕ = 1 / (r * sin(θ)) * dsrdr * dylmdϕ(l, m, θ, ϕ)
         return SVector(ur, uθ, uϕ)
     else
-        sc = s(Basis{T}, V, l, m, n, V.r1)
+        sc = s(Basis{T, Vol}, V, l, m, n, V.r1)
         ur = l * sc * ylm(l, m, θ, ϕ) * (l + 1) * r^(-l - 2)
         uθ = -l / r * sc * r^(-l - 1) * dylmdθ(l, m, θ, ϕ)
         uϕ = -l / (r * sin(θ)) * r^(-l -1) * sc * dylmdϕ(l, m, θ, ϕ)
@@ -28,11 +28,11 @@ function poloidal_discretize(::Type{Basis{T}}, V::Volume, l, m, n, r, θ, ϕ) wh
 end
 
 
-function toroidal_discretize(::Type{Basis{T}}, V::Volume, l, m, n, r, θ, ϕ) where {T}
+function toroidal_discretize(::Type{Basis{T, Vol}}, V::Vol, l, m, n, r, θ, ϕ) where {T, Vol<:Volume}
     if r <= V.r1
         ur = 0.0
-        uθ = 1 / sin(θ) * t(Basis{T}, V, l, m, n, r) * dylmdϕ(l, m, θ, ϕ)
-        uϕ = -t(Basis{T}, V, l, m, n, r) * dylmdθ(l, m, θ, ϕ)
+        uθ = 1 / sin(θ) * t(Basis{T, Vol}, V, l, m, n, r) * dylmdϕ(l, m, θ, ϕ)
+        uϕ = -t(Basis{T, Vol}, V, l, m, n, r) * dylmdθ(l, m, θ, ϕ)
         return SVector(ur, uθ, uϕ)
     else
         return SVector(0.0, 0.0, 0.0)
