### 2025-04-19 b9b01d6
up docs manifest
