+++ src/Discretization.jl
@@ -280,24 +280,24 @@ function coeffs_to_SHTnSlmTlm!(Qlmu, Slmu, Tlmu, Qlmb, Slmb, Tlmb, lmnpu, lmntu,
 
 	for (c,(l,m,n)) in zip(@views(coeffs[1:npu]),lmnpu)
 		@inline p = r-> s(TU,u.V,l,m,n,r)
-		Qlmu[SHTns.LM_cplx(sht,l,m)] += c*l*(l+1)/r*p(r)
-		Slmu[SHTns.LM_cplx(sht,l,m)] += c/r*∂(r->r*p(r),r)
+		Qlmu[SHTns.LM(sht,l,m)] += c*l*(l+1)/r*p(r)
+		Slmu[SHTns.LM(sht,l,m)] += c/r*∂(r->r*p(r),r)
 
 	end
 	for (c,(l,m,n)) in zip(@views(coeffs[npu+1:npu+ntu]),lmntu)
-		Tlmu[SHTns.LM_cplx(sht,l,m)] += c*t(TU,u.V,l,m,n,r)
+		Tlmu[SHTns.LM(sht,l,m)] += c*t(TU,u.V,l,m,n,r)
 	end
 
 	#magnetic field
 
 	for (c,(l,m,n)) in zip(@views(coeffs[nu+1:nu+npb]),lmnpb)
 		@inline p = r-> s(TB,b.V,l,m,n,r)
-		Qlmb[SHTns.LM_cplx(sht,l,m)] += c*l*(l+1)/r*p(r)
-		Slmb[SHTns.LM_cplx(sht,l,m)] += c/r*∂(@inline(r->r*p(r)),r)
+		Qlmb[SHTns.LM(sht,l,m)] += c*l*(l+1)/r*p(r)
+		Slmb[SHTns.LM(sht,l,m)] += c/r*∂(@inline(r->r*p(r)),r)
 
 	end
 	for (c,(l,m,n)) in zip(@views(coeffs[nu+npb+1:nu+npb+ntb]),lmntb)
-		Tlmb[SHTns.LM_cplx(sht,l,m)] += c*t(TB,b.V,l,m,n,r)
+		Tlmb[SHTns.LM(sht,l,m)] += c*t(TB,b.V,l,m,n,r)
 	end
 
 	return nothing
@@ -315,12 +315,12 @@ function coeffs_to_SHTnSlmTlm!(Qlmu, Slmu, Tlmu, lmnpu, lmntu, coeffs, sht, u::T
 
 	for (c,(l,m,n)) in zip(@views(coeffs[1:npu]),lmnpu)
 		@inline p = r-> s(TU,u.V,l,m,n,r)
-		Qlmu[SHTns.LM_cplx(sht,l,m)] += c*l*(l+1)/r*p(r)
-		Slmu[SHTns.LM_cplx(sht,l,m)] += c/r*∂(r->r*p(r),r)
+		Qlmu[SHTns.LM(sht,l,m)] += c*l*(l+1)/r*p(r)
+		Slmu[SHTns.LM(sht,l,m)] += c/r*∂(r->r*p(r),r)
 
 	end
 	for (c,(l,m,n)) in zip(@views(coeffs[npu+1:npu+ntu]),lmntu)
-		Tlmu[SHTns.LM_cplx(sht,l,m)] += c*t(TU,u.V,l,m,n,r)
+		Tlmu[SHTns.LM(sht,l,m)] += c*t(TU,u.V,l,m,n,r)
 	end
 
 	return nothing
@@ -330,7 +330,7 @@ end
 function discretizationsetup(L::Int, V::Volume)
 	
 	nr = 2L
-	sht = SHTnsCfg(L)
[…]
