### 2025-04-20 fgerick/Limace.jl#20 @jakelangham
[JOSS Review] README.md Malkus modes example
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7593.

I installed Limace today on two systems with Julia 1.11.5 and tried the examples on the README.md page in order.

On both systems, the following step fails:

`RHSi = Limace.induction(b,u,B0)`

Unfortunately, I'm not yet familiar enough with the package to say anything intelligent about why this is, but here is the error message:

> KeyError: key (1, -1, 0) not found
> Stacktrace:
>  [1] getindex
>    @ ./dict.jl:477 [inlined]
>  [2] #_crossterm!#197
>    @ ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:261 [inlined]
>  [3] _crossterm!
>    @ ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:252 [inlined]
>  [4] _induction(::Val{false}, bbi::Basis{PerfectlyConducting, Limace.Bases.Sphere}, buj::Basis{Inviscid, Limace.Bases.Sphere}, B0::BasisElement{Basis{PerfectlyConducting, Limace.Bases.Sphere}, Toroidal, Float64}; external::Bool)
>    @ Limace ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:425
>  [5] _induction
>    @ ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:404 [inlined]
>  [6] induction(bbi::Basis{PerfectlyConducting, Limace.Bases.Sphere}, buj::Basis{Inviscid, Limace.Bases.Sphere}, B0::BasisElement{Basis{PerfectlyConducting, Limace.Bases.Sphere}, Toroidal, Float64}; threads::Bool, external::Bool)
>    @ Limace ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:801
>  [7] induction(bbi::Basis{PerfectlyConducting, Limace.Bases.Sphere}, buj::Basis{Inviscid, Limace.Bases.Sphere}, B0::BasisElement{Basis{PerfectlyConducting, Limace.Bases.Sphere}, Toroidal, Float64})
>    @ Limace ~/.julia/packages/Limace/RPZHN/src/forces/induction.jl:801
>  [8] top-level scope
>    @ REPL[11]:1

However, I suspect the error is the line

`u = Inviscid(N; m=1)`

since you say beforehand "Without specifying the azimuthal wave number", but `m=1` appears to do exactly this. If I run instead

`u = Inviscid(N)`

there is no error.

Additionally, for the next line

`RHSd = spzeros(length(b),length(u)) #empty (no Ohmic diffusion)`

to work, it seems the user needs `using SparseArrays` beforehand, so this should be added above somewhere.
