### 2025-04-22 fgerick/Limace.jl#20 @fgerick
I have updated this. In order to work when executing all code cells! see 1887073
