+++ src/forces/specializations.jl
@@ -2,7 +2,7 @@ const Tunion = Union{Insulating, Inviscid, Viscous, ThinWall} #all bases that ca
 
 # only assemble cross terms that are neigboring up to max(ni-(n0+1)-l0,first(njs_all)):min(ni+n0+1+l0,last(njs_all))
 
-@inline function _crossterm!(bbi::Basis{Ti,V}, U0::BasisElement{Basis{T0},PT,T}, bbj::Basis{Tj,V}, is, js, aijs, i0, j0,
+@inline function _crossterm!(bbi::Basis{Ti,V}, U0::BasisElement{Basis{T0,V},PT,T}, bbj::Basis{Tj,V}, is, js, aijs, i0, j0,
     li, mi, lj, mj, rwrs, lmn2k_bi, lmn2k_bj, nrangefi, nrangefj, indf, EA; kwargs...) where {Ti<:Tunion, Tj<:Tunion, T0<:Tunion, PT<:Helmholtz,T, V<:Volume}
     l0,m0,n0 = U0.lmn
     for ni in nrangefi(bbi, li)
@@ -18,7 +18,7 @@ const Tunion = Union{Insulating, Inviscid, Viscous, ThinWall} #all bases that ca
     return nothing
 end
 
-@inline function _crossterm!(bbi::Basis{Ti,V}, buj::Basis{Tj,V}, B0::BasisElement{Basis{T0},PT,T}, is, js, aijs, i0, j0,
+@inline function _crossterm!(bbi::Basis{Ti,V}, buj::Basis{Tj,V}, B0::BasisElement{Basis{T0,V},PT,T}, is, js, aijs, i0, j0,
     li, mi, lj, mj, rwrs, lmn2k_bi, lmn2k_uj, nrangefi, nrangefj, indf, EA; kwargs...) where {Ti<:Tunion, Tj<:Tunion, T0<:Tunion, PT<:Helmholtz,T, V<:Volume}
     l0,m0,n0 = B0.lmn
     for ni in nrangefi(bbi, li)
