### 2025-04-26 fgerick/Limace.jl#21 @jakelangham
[JOSS Review] paper suggestions
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7593.

I've read through the paper and think it's well written and appropriate for the journal etc. I have a various minor suggestions to raise below, before I check this part of my review off.

- line 20: typo, change 'astrophisical' -> 'astrophysical'
- line 21: typo, change 'mode' -> 'modes'
- linearized equations: Since there is not room to develop these fully, consider including a reference early on, to make it easy for people to look up the detail. Also, I am somewhat pedantically against calling them 'the linearized equations', since really there are two steps combined here - the equations are linearised to get du/dt = ... and db/dt = ..., and then decomposed into separate equations for each mode. So maybe you could allude to the modal decomposition to be extra clear, or say 'linearized equations for each mode' or something.
- When you say 'One of the only open-source models ... Kore', it makes it seem like you have left out some. So you do actually mean Kore is the _only_ open-source code that computes these modes (at least in physically relevant regimes)? Also, it was not quite clear to me if you are trying to say that Kore is difficult to set up, or if both Kore and Limace are trying to solve the same essential problem - providing a high-level interface for these computations - but with a slightly different focus / features.
- line 38: The vector $\mathbf{\Omega}$ doesn't seem to be defined - it's only a scalar quantity in the equations. Likewise, $P$ does not appear in the equations, only $p$.
- It would be good to (briefly) define the spherical coordinate system somewhere. I see $\theta$, $\phi$ and $r$ being used, but then confusingly, $\mathbf{e}_z$ and $\mathbf{r}$ as basis vectors.
- line 42: You could say '...projected onto trial vectors $\mathbf{u}_i$ and $\mathbf{b}_i$' here, so people know that's what $\mathbf{u}_i$ and $\mathbf{b}_i$ are in the forthcoming eqs.
- line 46: I'm not familiar with this square bracket notation ([P, Q] etc), at least apart from things like Poisson brackets. Is it a standard way to write this?
- line 49: I understand why you want to use $P$ here, but this conflicts with the pressure.
- line 51: The selection of boundary conditions seems very important to me. Perhaps you could list at least the main ones that Limace supports?
- line 62: 'eigen problem' (and later 'eigen solution') would usually be just one word I think.
- line 67: 'To compute few eigen solutions of the sparse system...' I don't fully understand this sentence, so perhaps rephrase? Maybe it should be something like 'For larger systems where memory requirements render this infeasible, the leading eigensolutions can be computed by ...'
