### 2025-05-05 2c87129
paper suggestions issue 21
