+++ paper/paper.md
@@ -17,7 +17,7 @@ affiliations:
  - name: Royal Observatory of Belgium, Belgium
    index: 2
    ror: 00hjks330
-date: 7 October 2024
+date: 5 May 2024
 bibliography: paper.bib
 ---
 
@@ -34,31 +34,30 @@ whilst remaining computationally performant enough to tackle relevant physical p
 The study of hydromagnetic modes is relevant in particular to Earth's liquid core. 
 Despite having been theoretically predicted a long time ago [@hidefree1966; @malkushydromagnetic1967; @braginskytorsional1970], 
 recent advances in numerical modelling and new observational evidence in geomagnetic data have reignited interest in these modes [@gerickfast2021; @gilletsatellite2022; @trianacore2022; @luowaves2022a; @luowaves2022; @gerickinterannual2024].
-It is therefore relevant to the geophysical and astrophisical fluid dynamics community to have access to a code that models these mode.
-One of the only open-source models to compute hydromagnetic modes in relevant parameters for planetary cores is Kore [@trianaviscous2021], 
+It is therefore relevant to the geophysical and astrophysical fluid dynamics community to have access to a code that models these modes.
+The only open-source models to compute hydromagnetic modes in relevant parameters for planetary cores is Kore [@trianaviscous2021], 
 a spectral code based on ultraspherical polynomials in axisymmetric setups written in Python.
-The implementation of such a code requires substantial work due to the complexity of spectral equations that govern the modes.
-This is true also for very basic and idealized examples, resulting in a substantial entry barrier for scientists to model these modes.
-`Limace.jl` tries to lower this entry barrier, by providing an open source model with a very simple high-level API and modern online documentation with practical examples.
+Although Kore is limited to axisymmetric geometries, the implementation requires substantial work due to the complexity of the underlying spectral equations, resulting in a substantial entry barrier for scientists to model these modes.
+In addition to Kore, `Limace.jl` tries to lower this entry barrier, by providing a generic open source model with a very simple high-level API and modern online documentation with practical examples.
 
 Despite having a high-level interface, `Limace.jl` can be used to solve complex and geophysically relevant problems.
-A unique feature of `Limace.jl` is the support of complex background magnetic fields and flows over which the modes evolve.
+A unique feature of `Limace.jl` is the support of complex (non-axisymmetric) background magnetic fields and flows over which the modes evolve.
 The code has been developed from the beginning to leave assumptions of symmetry up to the user.
 The model code base is tested against mode solutions from the scientific literature to ensure its correctness.
 
 # Theoretical background and implementation details
 
-In order to compute modal solutions, we consider the linearized momentum equation of the incompressible fluid and the linearized induction equation
+In order to compute modal solutions, we consider the linearized momentum equation of the incompressible fluid and the linearized induction equation [@iversscalar2008, @gerickinterannual2024]
 $$
 \begin{aligned}
[…]
