### 2025-05-05 fgerick/Limace.jl#21 @fgerick
- [x] line 20: typo, change 'astrophisical' -> 'astrophysical'

- [x] line 21: typo, change 'mode' -> 'modes'
- [x] linearized equations: Since there is not room to develop these fully, consider including a reference early on, to make it easy for people to look up the detail. Also, I am somewhat pedantically against calling them 'the linearized equations', since really there are two steps combined here - the equations are linearised to get du/dt = ... and db/dt = ..., and then decomposed into separate equations for each mode. So maybe you could allude to the modal decomposition to be extra clear, or say 'linearized equations for each mode' or something.

I have cited the relevant papers before the equations. The equations are not linearized equations for each mode. What is done is the linearization and then a further decomposition of the velocity and magnetic field bases elements is provided in terms of poloidal and toroidal vectors. I have changed the wording to highlight that these are not equations for one mode.

- [x] When you say 'One of the only open-source models ... Kore', it makes it seem like you have left out some. So you do actually mean Kore is the _only_ open-source code that computes these modes (at least in physically relevant regimes)? Also, it was not quite clear to me if you are trying to say that Kore is difficult to set up, or if both Kore and Limace are trying to solve the same essential problem - providing a high-level interface for these computations - but with a slightly different focus / features.

I have changed the text to try to be more precise. It

- [x] line 38: The vector $\mathbf{\Omega}$ doesn't seem to be defined - it's only a scalar quantity in the equations. Likewise, $P$ does not appear in the equations, only $p$.

- [x] It would be good to (briefly) define the spherical coordinate system somewhere. I see $\theta$, $\phi$ and $r$ being used, but then confusingly, $\mathbf{e}_z$ and $\mathbf{r}$ as basis vectors.

I have removed $\mathbf{e}_z$ to avoid introducing cylindrical coordinates also.

- [x] line 42: You could say '...projected onto trial vectors $\mathbf{u}_i$ and $\mathbf{b}_i$' here, so people know that's what $\mathbf{u}_i$ and $\mathbf{b}_i$ are in the forthcoming eqs.
- [x] line 46: I'm not familiar with this square bracket notation ([P, Q] etc), at least apart from things like Poisson brackets. Is it a standard way to write this?

It has been used a few times in the fluid dynamics community to avoid dublicate equations, but I have removed it here and just state the equations for P and Q and refer to S and T as analogously.

- [x] line 49: I understand why you want to use $P$ here, but this conflicts with the pressure.
- [x] line 51: The selection of boundary conditions seems very important to me. Perhaps you could list at least the main ones that Limace supports?
- [x] line 62: 'eigen problem' (and later 'eigen solution') would usually be just one word I think.
- [x] line 67: 'To compute few eigen solutions of the sparse system...' I don't fully understand this sentence, so perhaps rephrase? Maybe it should be something like 'For larger systems where memory requirements render this infeasible, the leading eigensolutions can be computed by ...'
