### 2025-05-06 #7593 @fgerick
I have responded to the changes of @jakelangham to the PDF article [[JOSS Review] paper suggestions fgerick/Limace.jl#21](https://github.com/fgerick/Limace.jl/issues/21). Is there anything else I can do? Is the review of @milankl not relevant any longer?
