### 2025-05-25 fgerick/Limace.jl#22 @jakelangham
[JOSS Review] Documentation suggestions
This issue relates to ongoing reviews at https://github.com/openjournals/joss-reviews/issues/7593.

My impressions of reading the documentation are that what's there is pretty polished and useful. Nevertheless, I caught some typos and have various other minor suggestions, which I hope will be useful, listed below. I also have the following more significant point.

I am broadly concerned that when it comes to the 'API'-style documentation (Bases, Forces, Poly, Postprocessing), the information given is too sparse to be useful for a user to move beyond the examples. For example, the documentation of Bases, doesn't seem to fully reflect how bases are set up in the given use cases. Then, when it comes to Forces, each function is listed, but there is no independent guidance as to what the arguments mean (e.g. does `factor` in `Diffusion` set the viscosity?). This needn't be a big change, but I suggest going through each of these function definitions and making sure that it would be clear for a new user how to make use of them without having to read the source code, since this sort of thing is a requirement for JOSS acceptance.

**Quickstart.**

- Typo: 'Chose a polynomial truncation, e.g. N=10, and an appropriate bases' -> 'Chose a polynomial truncation, e.g. N=10, and an appropriate basis'
- Where you say 'The forcings considered in the problem are collected in a list to be included', I think it would be helpful to be a bit more explicit about what's going on here, i.e. that (as far as I understand...) the specified list is arbitrary and constructs the linear system in a modular way, allowing a great deal of flexibility in the problem being solved. You could also link to the 'Forces' section, as you did for 'Bases'.
- The line 'We can use a keyword `threads=true` to assemble the matrices using multithreading on a single computer (see [Limace.assemble!](https://fgerick.github.io/Limace.jl/dev/quickstart/#Limace.assemble!)).', because the documentation for `Limace.assemble!` below does not list any of the `kwargs`.

**Examples.**

- The Inviscid example is very clear and helpful. However, when testing that the solution constructed the first way is the same as produced by the high-level interface, the line `(λ ≈ λ2) && (x ≈ x2)` fails. This seems to be because complex conjugate modes get reordered depending on the solution method. So, e.g. `sort(imag(λ)) ≈ sort(imag(λ2))` reports `true`, but I'm sure you can think of a cleverer way to correct this test.
- In the 'Visualisation' part of the inviscid example, the `spectospat` is playing a big role, but hasn't been introduced yet. It would be good to introduce or just say 'see Postprocessing' or similar.
- Typo: 'Makus modes' -> 'Malkus modes'
- In 'Malkus modes', it's not so clear what $s$ is set to when the problem is solved.
- In the 'Alfven modes' Assembly step, the numbers 0.3 and 0.7 are... the strength of the background fields? Why these choices?
- In 'Alfven modes', where you say 'The keyword nev can be set to the desired number of eigenvalues.' it would be good to move this higher up since `nev` is also used when calling `Limace.solve!`.

**Theoretical background.**

- In (5), replace the trailing '.' with a ','
- In (6), I suppose that $\Omega e_z$ should be $\mathbf{\Omega}$ for consistency with the equations above and the writeup in the paper. This also applies elsewhere in the documentation.

**Bases.**

- It is not clear to me without reading between the lines, how the basis functions `s` and `t` relate to the writeup in the Theoretical Background. This underpins the whole section so consider defining up front.
- Likewise, the functions $J_n$ are not defined, which is confusing because $J_n$ in my mind normally denotes a Bessel function.
- In the text for `Limace.InsulatingBasis.unitspherenorm` there is a typo causing the first integral to render incorrectly.

**Poly.**

- A typo in the definition of `Limace.Poly.jacobi` means the mathematics isn't rendering.

**Solving the eigenvalue problem.**

- In 'Sparse eigenvalue solver' - perhaps it's confusing to use $\sigma$ for the eigenvalue shift, since in the Theoretical Background, $\sigma$ was the real part of $\lambda$.

**Postprocessing.**

- In `Limace.Processing.lmn_n` - possible typesetting error?

(N.B. Don't feel obliged to respond to all of the small issues individually!)
