### 2025-05-26 fgerick/Limace.jl#13 @milankl
> I have not managed to get a useful version of ProgressMeter.jl running.

Yeah I can imagine that this isn't easy in the case of loop iterations being varingly costly which itself is, I guess, hard to predict. Closing this issue as it was more a feature request and so I'm happy you looked into it and reasonably decided it's not feasible right now.
