### 2025-05-26 fgerick/Limace.jl#15 @milankl
> This is a big complicated task

I know. For bigger models that can often take over any other efforts to implement features. I'm happy now to see that the initial building blocks are there, you document them and it sounds like you have it on your radar which isn't an easy issue to tick off but a journey. There's many ways to make things more modular and the details of how to do this are complicated and change over time. Happy for this issue to be closed!
