### 2025-05-26 fgerick/Limace.jl#17 @milankl
Closing this as you addressed all my points and I'm very happy with the direction this package went! Well done. I'll finish the review over in the JOSS issue.
