### 2025-05-27 15b8a1e
add cartesian degree determination
