### 2025-05-27 5e289f5
theory, bases, poly, eigen, processing small comments
