### 2025-05-27 ca5911c
change lmn n docstring
