+++ docs/src/bases.md
@@ -9,25 +9,67 @@ The function arguments of `s` and `t` include the azimuthal degree `m` to keep i
 
 Currently, all bases rely on a radial representation using Jacobi polynomials $J_n^{\alpha,\beta}(x)$ (implemented as [Limace.Poly.jacobi](@ref)).
 
+All implemented bases are of type `Basis{T,Vol<:Volume}`, where the type `T` defines the type of the basis (e.g. `Inviscid`).
+```@docs
+Limace.Bases.Basis
+```
 
 ## Inviscid velocity basis
 
 For an inviscid fluid, we only require ``\mathbf{u}\cdot\mathbf{n}=0`` at the boundary. This is known as the non-penetration condition. For poloidal and toroidal decomposition, this boils down to the requirement that the poloidal scalar vanishes at the boundary. The toroidal scalar remains unconstrained and we have free slip.
 
+The basis is orthonormal on the unit sphere $\mathcal{V}$, so that
+```math
+\int_\mathcal{V} \mathbf{u}_i^*\cdot\mathbf{u}_j\,\mathrm{d}V = \delta_{ij}.
+```
+
 ```@autodocs
 Modules = [Limace.InviscidBasis]
+```
+
+
+To define the inviscid basis, using a polynomial truncation `N = 10`, simply run
+```@example
+using Limace #hide
+N = 10
+u = Inviscid(N)
+```
+
+The basis `u` will include all azimuthal wave numbers `m`. To restrict to a single wave number (e.g. `m=1`), one can do
+
+```@example
+using Limace #hide
+N = 10 #hide
+u = Inviscid(N; m=1)
 ```
 
 ## Viscous velocity basis 
 
 For a viscous fluid, in addition to the non-penetration condition, we require no-slip, so that in total ``\mathbf{u} = \mathbf{0}`` at the boundary. [chen_optimal_2018](@citet) have written a basis that satisfies these conditions, and requires orthogonality w.r.t the vector Laplacian, so that
 ```math
-\int \mathbf{u}_i^*\cdot\boldsymbol{\nabla}^2\mathbf{u}_j\,\mathrm{d}V = \delta_{ij}.
+\int_\mathcal{V} \mathbf{u}_i^*\cdot\boldsymbol{\nabla}^2\mathbf{u}_j\,\mathrm{d}V = \delta_{ij}.
 ```
 This is desirable, as the resulting projections of all considered forces are banded (if only combined with another basis that satisfies the appropriate orthogonality).
 
 ```@autodocs
 Modules = [Limace.ViscousBasis]
+```
+
+
+
+To define the viscous basis, using a polynomial truncation `N = 10`, simply run
+```@example
+using Limace #hide
+N = 10
+u = Viscous(N)
+```
+
+The basis `u` will include all azimuthal wave numbers `m`. To restrict to a single wave number (e.g. `m=1`), one can do
+
+```@example
+using Limace #hide
+N = 10 #hide
+u = Viscous(N; m=1)
 ```
 
 ## Insulating magnetic field basis
@@ -48,8 +90,28 @@ t_{lmn}\bigg|_{\partial\mathcal{V}} &= 0.
 ```
 for all $l,m,n$.
 
+The basis is orthonormal w.r.t the vector Laplacian, so that
+```math
+\int_{\mathbb{R}^3} \mathbf{B}_i^*\cdot\boldsymbol{\nabla}^2\mathbf{B}_j\,\mathrm{d}V = \delta_{ij}.
+```
+
 ```@autodocs
 Modules = [Limace.InsulatingBasis]
+```
+
+To define the insulating basis, using a polynomial truncation `N = 10`, simply run
+```@example
+using Limace #hide
+N = 10
+b = Insulating(N)
+```
+
[…]
