### 2025-05-28 0e48ea9
improve poly documentation and add quadrature
