### 2025-05-28 75772c6
clean up induction crossterms
