+++ src/forces/induction.jl
@@ -343,91 +343,6 @@ Fallback functions for `_crossterm!` term for `B0`. Write specialized function t
     return nothing
 end
 
-#∫Sᵢ⋅∇×(pⱼ×S₀) dV
-function _induction_sps!(bbi::TI, buj::TJ, B0::BasisElement{T0,Poloidal,T}, is, js, aijs, i0, j0,
-    li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_uj, args...; external=true) where {TI<:Basis,TJ<:Basis,T0<:Basis,T}
-    return _crossterm!(bbi, buj, B0, is, js, aijs, i0, j0, li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_uj, nrange_p_bc, nrange_p, _induction_sSS, args...; external)
-end
-
-#∫Sᵢ⋅∇×(p₀×Sⱼ) dV
-function _induction_sps!(bbi::TI, U0::BasisElement{T0,Poloidal,T}, bbj::TJ, is, js, aijs, i0, j0,
-    li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_bj, args...; external=true) where {TI<:Basis,TJ<:Basis,T0<:Basis,T}
-    return _crossterm!(bbi, U0, bbj, is, js, aijs, i0, j0, li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_bj, nrange_p_bc, nrange_p, _induction_sSS, args...; external)
-end
-
-#∫Sᵢ⋅∇×(pⱼ×T₀) dV
-function _induction_spt!(bbi::TI, buj::TJ, B0::BasisElement{T0,Toroidal,T}, is, js, aijs, i0, j0,
-    li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_uj, args...) where {TI<:Basis,TJ<:Basis,T0<:Basis,T}
-    return _crossterm!(bbi, buj, B0, is, js, aijs, i0, j0, li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_p_uj, nrange_p_bc, nrange_p, _induction_sTS, args...)
-end
-
-#∫Sᵢ⋅∇×(p₀×Tⱼ) dV
-function _induction_spt!(bbi::TI, U0::BasisElement{T0,Poloidal,T}, bbj::TJ, is, js, aijs, i0, j0,
-    li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_t_bj, args...) where {TI<:Basis,TJ<:Basis,T0<:Basis,T}
-    return _crossterm!(bbi, U0, bbj, is, js, aijs, i0, j0, li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_t_bj, nrange_p_bc, nrange_t, _induction_sTS, args...)
-end
-
-#∫Sᵢ⋅∇×(qⱼ×S₀) dV
-function _induction_sqs!(bbi::TI, buj::TJ, B0::BasisElement{T0,Poloidal,T}, is, js, aijs, i0, j0,
-    li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_t_uj, args...; external=true) where {TI<:Basis,TJ<:Basis,T0<:Basis,T}
-    return _crossterm!(bbi, buj, B0, is, js, aijs, i0, j0, li, mi, lj, mj, rwrs, lmn2k_p_bi, lmn2k_t_uj, nrange_p_bc, nrange_t, _induction_tSS, args...; external)
-end
-
-#∫Sᵢ⋅∇×(q₀×Sⱼ) dV
[…]
