+++ docs/src/postprocessing.md
@@ -5,25 +5,147 @@ A non-exhaustive collection of spectral to spatial discretization and postproces
 
 ## Spatial discretization
 
-In the `Limace.Discretization` submodule, some methods to discretize `BasisElement` and eigenvectors are provided.
+In the `Limace.Discretization` submodule, some methods to discretize [BasisElement](@ref) and eigenvectors are provided.
 
-```@autodocs
-Modules = [Limace.Discretization]
+A fully manual discretization routine is given by
+```@docs
+Limace.Discretization.discretize
+```
+To evaluate one [BasisElement](@ref) at a given point $(r,\theta,\phi)$, we can do
+```@example processing
+using Limace #hide
+using Limace.Discretization: discretize
+u = Inviscid(10)
+b = BasisElement(u, Poloidal, (2,0,1))
+r,θ,ϕ = 0.9, π/4, 3π/2
+discretize(b, r,θ,ϕ)
+```
+
+The same also works for a vector of [BasisElement](@ref)s.
+```@example processing
+bs = [BasisElement(u, Poloidal, (2,0,1)), BasisElement(u, Poloidal, (2,0,2))]
+discretize(bs, r,θ,ϕ)
+```
+
+To discretize an eigenvector using [Limace.Discretization.discretize](@ref), consider the following example of the Malkus modes
+```@example processing
+N = 6
+Le = 1e-1
+
+u = Inviscid(N)
+b = PerfectlyConducting(N)
+
+bases = [u,b]
+
+B0 = BasisElement(b, Toroidal, (1,0,0), 2sqrt(2pi/15))
+forcings = [Limace.Inertial(u), Limace.Inertial(b), Limace.Coriolis(u, 1/Le), Limace.Lorentz(u, b, B0), Limace.InductionB0(b,u,B0)]
+
+problem = LimaceProblem(bases, forcings)
+Limace.assemble!(problem)
+Limace.solve!(problem)
+nothing #hide
+```
+
+We can chose a single eigenvector `x` and define some arbitrary spatial grid for `r`, `θ` and `ϕ`:
+```@example processing
+x = problem.sol.vectors[:,1]
+nr,nθ,nϕ = 20,30,40
+r = range(0.01,0.99, length=nr)
+θ = range(0.01,π-0.01,length=nθ)
+ϕ = range(0,2π, length=nϕ)
+ur,uθ,uϕ, br, bθ, bϕ = discretize(x, u, b, r, π/2 .- θ, ϕ)
+size(ur)
+```
+
+For more performant transforms to spatial space, it is recommended to use [Limace.Discretization.spectospat](@ref), 
+which relies on fast vector spherical harmonic transforms provided through [SHTns.jl](https://github.com/fgerick/SHTns.jl).
+```@docs
+Limace.Discretization.spectospat
+```
+
+To discretize the eigenvector `x` on a Gauss-Legendre grid in `r` and `θ`, as well as equidistant grid in `ϕ`, we can call `spectospat`:
+```@example processing
+using Limace.Discretization: spectospat
+ur,uθ,uϕ, br,bθ,bϕ, r,θ,ϕ = spectospat(x, u, b, nr, nθ, nϕ)
+nothing #hide
 ```
 
 ## Filtering of spectrum
 
+When the background state couples different spectral degrees (may be in radius, spherical harmonic degree or order), the solutions are only truncated approximations of the true solution. 
+This is not the case, for example, for the [Inviscid inertial modes](@ref "Inviscid inertial modes in the sphere") or the [Malkus modes](@ref "Malkus modes in the sphere"), which are fully resolved at one truncation degree and are therefore perfectly converged.
[…]
