### 2025-06-02 #7593 @fgerick
Thank you very much @milankl and @jakelangham for the instructive comments that really helped a lot in improving the manuscript, code base and its documentation. I have taken into account the latest comments of @jakelangham to improve the documentation. Please let me know if there are any other things to improve!
