+++ dev/refs.bib
@@ -95,7 +95,7 @@
   journal = {Proceedings of the Royal Society of London. Series A, Mathematical and Physical Sciences},
   volume = {331},
   number = {1587},
-  eprint = {78350},
+  doi = {10.1098/rspa.1973.0003},
   eprinttype = {jstor},
   pages = {469--478},
   publisher = {The Royal Society},
