### 2025-06-10 a1b119b
rm perfectlyconducting in inviscid
