+++ v0.1.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"Braginsky, S. I. (1970). Torsional Magnetohydrodynamics Vibrations in the Earth's Core and Variations in Day Length. Geomagn. Aeron. 10, 3–12.\n\n\n\nBullard, E. C. and Gellman, H. (1954). Homogeneous Dynamos and Terrestrial Magnetism. Philosophical Transactions of the Royal Society of London. Series A, Mathematical and Physical Sciences 247, 213–278.\n\n\n\nChen, L.; Herreman, W.; Li, K.; Livermore, P. W.; Luo, J. W. and Jackson, A. (2018). The Optimal Kinematic Dynamo Driven by Steady Flows in a Sphere. Journal of Fluid Mechanics 839, 1–32.\n\n\n\nGerick, F. and Livermore, P. W. (2024). Interannual Magneto–Coriolis Modes and Their Sensitivity on the Magnetic Field within the Earth's Core. Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences 480, 20240184.\n\n\n\nIvers, D. J. and Phillips, C. G. (2008). Scalar and Vector Spherical Harmonic Spectral Equations of Rotating Magnetohydrodynamics. Geophysical Journal International 175, 955–974.\n\n\n\nJames, R. W. (1973). The Adams and Elsasser Dynamo Integrals. Proceedings of the Royal Society of London. Series A, Mathematical and Physical Sciences 331, 469–478.\n\n\n\nLivermore, P. W. (2014). A Compendium of Galerkin Orthogonal Polynomials.\n\n\n\nMalkus, W. V. (1967). Hydromagnetic Planetary Waves. Journal of Fluid Mechanics 28, 793–802.\n\n\n\nZhang, K.; Earnshaw, P.; Liao, X. and Busse, F. H. (2001). On Inertial Waves in a Rotating Fluid Sphere. Journal of Fluid Mechanics 437, 103–119.\n\n\n\n","category":"page"},{"location":"poly/#Poly","page":"Poly","title":"Poly","text":"","category":"section"},{"location":"poly/","page":"Poly","title":"Poly","text":"Polynomial functions and conveniences that are needed for the spectral or spatial discretization.","category":"page"},{"location":"poly/#Polynomials","page":"Poly","title":"Polynomials","text":"","category":"section"},{"location":"poly/","page":"Poly","title":"Poly","text":"Limace.Poly.jacobi\nLimace.Poly.ylm","category":"page"},{"location":"poly/#Limace.Poly.jacobi","page":"Poly","title":"Limace.Poly.jacobi","text":"jacobi(n,a,b,x)\n\nJacobi polynomial \n\nJ_n^(ab)(x)\n\n\n\n\n\n","category":"function"},{"location":"poly/#Limace.Poly.ylm","page":"Poly","title":"Limace.Poly.ylm","text":"ylm(ℓ::Int64, m::Int64, θ, φ) -> Any\n\n\nSpherical harmonic Y_l^m in full norm, i.e. \n\nint Y_l^mY_i^j sin(theta)mathrmdthetamathrmdphi = delta_lidelta_mj\n\nwhere theta is the colatitude and phi the azimuthal angle.\n\n\n\n\n\n","category":"function"},{"location":"poly/#Integral-functions","page":"Poly","title":"Integral functions","text":"","category":"section"},{"location":"poly/","page":"Poly","title":"Poly","text":"Integrals over the spherical surfaces are reduced to two specific integrals, known as the Adam-Gaunt and Elsasser integrals (James, 1973).","category":"pag
[…]
