+++ v0.1.0/refs.bib
@@ -1,0 +1,103 @@
+@article{ivers_scalar_2008,
+  title = {Scalar and Vector Spherical Harmonic Spectral Equations of Rotating Magnetohydrodynamics},
+  author = {Ivers, D. J. and Phillips, C. G.},
+  year = {2008},
+  month = dec,
+  journal = {Geophysical Journal International},
+  volume = {175},
+  number = {3},
+  pages = {955--974},
+  issn = {0956-540X},
+  doi = {10.1111/j.1365-246X.2008.03944.x}
+}
+
+@article{bullard_homogeneous_1954,
+  title = {Homogeneous Dynamos and Terrestrial Magnetism},
+  author = {Bullard, Edward Crisp and Gellman, H.},
+  year = {1954},
+  month = nov,
+  journal = {Philosophical Transactions of the Royal Society of London. Series A, Mathematical and Physical Sciences},
+  volume = {247},
+  number = {928},
+  pages = {213--278},
+  publisher = {Royal Society},
+  doi = {10.1098/rsta.1954.0018}
+}
+
+@article{gerick_interannual_2024,
+  title = {Interannual {{Magneto}}--{{Coriolis}} Modes and Their Sensitivity on the Magnetic Field within the {{Earth}}'s Core},
+  author = {Gerick, F. and Livermore, P. W.},
+  year = {2024},
+  month = oct,
+  journal = {Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
+  volume = {480},
+  number = {2299},
+  pages = {20240184},
+  publisher = {Royal Society},
+  doi = {10.1098/rspa.2024.0184}
+}
+
+@misc{livermore_compendium_2014,
+  title = {A {{Compendium}} of {{Galerkin Orthogonal Polynomials}}},
+  author = {Livermore, P. W.},
+  year = {2014},
+  url = {https://homepages.see.leeds.ac.uk/~earpwl/Galerkin/Galerkin.html}
+}
+
+@article{chen_optimal_2018,
+  title = {The Optimal Kinematic Dynamo Driven by Steady Flows in a Sphere},
+  author = {Chen, L. and Herreman, W. and Li, K. and Livermore, P. W. and Luo, J. W. and Jackson, A.},
+  year = {2018},
+  month = mar,
+  journal = {Journal of Fluid Mechanics},
+  volume = {839},
+  pages = {1--32},
+  publisher = {Cambridge University Press},
+  issn = {0022-1120, 1469-7645},
+  doi = {10.1017/jfm.2017.924},
+  langid = {english}
+}
+
+@article{zhang_inertial_2001,
+  title = {On Inertial Waves in a Rotating Fluid Sphere},
+  author = {Zhang, Keke and Earnshaw, Paul and Liao, Xinhao and Busse, F H},
+  year = {2001},
+  journal = {Journal of Fluid Mechanics},
+  volume = {437},
+  pages = {103--119},
+  doi = {10.1017/S0022112001004049}
+}
+
+@article{braginsky_torsional_1970,
+  title = {Torsional Magnetohydrodynamics Vibrations in the {{Earth}}'s Core and Variations in Day Length},
+  author = {Braginsky, S. I.},
+  year = {1970},
+  journal = {Geomagn. Aeron.},
+  volume = {10},
+  pages = {3--12}
+}
+
+@article{malkus_hydromagnetic_1967,
+  title = {Hydromagnetic Planetary Waves},
+  author = {Malkus, Willem V R},
+  year = {1967},
+  journal = {Journal of Fluid Mechanics},
+  volume = {28},
+  number = {4},
+  pages = {793--802},
+  doi = {10.1017/S0022112067002447}
+}
+
[…]
