### 2025-06-10 #7593 @fgerick
Thank you @diehlpk . I have added my current affiliation to the article, made a version release including the latest changes and archived it on zenodo (https://doi.org/10.5281/zenodo.15632097). The archive author details and the license match the JOSS paper. Let me know if what I have done is correct. I have not put the `joss` branch that includes the paper into the release, but I believe this is okay with JOSS policy?
