### 2025-06-10 #7593 @diehlpk
@fgerick

Make sure that the license listed for the archive is the same as the software license.

Could you please check on that? No new release is required and you can just add the software license.
