+++ paper/paper.bib
@@ -187,17 +187,6 @@ url = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
   doi = {10.1098/rsta.1966.0026}
 }
 
-@article{malkushydromagnetic1967,
-  title = {Hydromagnetic Planetary Waves},
-  author = {Malkus, Willem V R},
-  year = {1967},
-  journal = {Journal of Fluid Mechanics},
-  volume = {28},
-  number = {4},
-  pages = {793--802},
-  doi = {10.1017/S0022112067002447}
-}
-
 @article{braginskytorsional1970,
   title = {Torsional Magnetohydrodynamics Vibrations in the {{Earth}}'s Core and Variations in Day Length},
   author = {Braginsky, S. I.},
