### 2025-06-10 #7593 @diehlpk
> > ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/15565610102?check_suite_focus=true). The generated XML metadata file is invalid.
> > ID ref-malkushydromagnetic1967 already defined
>
> [@diehlpk](https://github.com/diehlpk) Sorry I had two times the reference in the bibfile, it is fixed on the latest joss branch.

No worries, I am trying it again.
