### 2025-06-18 #7593 @kthyng
@fgerick I see you are registered as a JOSS reviewer — thank you! It would be super helpful if you input some information about what language and topics you are able to review: https://reviewers.joss.theoj.org/reviewers
