### 2025-06-19 #7593 @fgerick
Thank you @kthyng and thank you everyone @diehlpk , @milankl and @jakelangham for the help and suggestions on the article. It was my first JOSS publication and I learned a lot!
